move new command to existing entry point for simplicity
